requirements.txt
pybind11
stanford_stk==0.7.0

torch==2.1.2

transformers==4.40.0
vllm==0.4.0.post1
xformers==0.0.23.post1
megablocks==0.5.1

torch_tb_profiler
matplotlib

compressed_tensors
accelerate==0.27.2